Add garage simulator (#11)
* Add GarageSimulator
* SimulationView is not created by GarageSimulator
* Refactor SimulatorView code into GarageSimulator
